Handle expansion options in list contents API
